BJS2-66383 made a merger of two RedisConfig
