+++ src/main/java/school/faang/user_service/config/RedisConfig.java
@@ -1,17 +1,24 @@
 package school.faang.user_service.config;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
+import school.faang.user_service.service.UserService;
 
 @Configuration
 @RequiredArgsConstructor
+@Slf4j
 public class RedisConfig {
 
     @Value("${spring.data.redis.host}")
@@ -24,6 +31,25 @@ public class RedisConfig {
     public JedisConnectionFactory jedisConnectionFactory() {
         RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(redisHost, redisPort);
         return new JedisConnectionFactory(config);
+    }
+
+    @Bean
+    public RedisMessageListenerContainer listenerContainer(
+            RedisConnectionFactory factory,
+            MessageListenerAdapter adapter
+    ) {
+        log.info("Creating RedisMessageListenerContainer for topic 'user_ban'");
+        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+        container.setConnectionFactory(factory);
+        container.addMessageListener(adapter, new ChannelTopic("user_ban"));
+        return container;
+    }
+
+    @Bean
+    public MessageListenerAdapter messageListener(UserService userService) {
+        MessageListenerAdapter adapter = new MessageListenerAdapter(userService, "banUser");
+        adapter.setSerializer(new StringRedisSerializer());
+        return adapter;
     }
 
     @Bean
